Move table generation to `build.rs` file
